Enable Exodus geometry support
- Added Exodus support for the webviewer:
hereby, the webviewer accounts for multiple sections
ending with "GEOMETRY" for different fields -
all element blocks
are also taken into account when enhancing the
discretization. As restricted by 4C, we only allow for
a single file in all geometry sections.
- Removed these geometry sections from the general
sections within the webviewer, since these are not meant to be
visualized.
- Absolute paths for the exodus files are necessary
due to VFileInput not knowing the local path of the
file.