Bugfix: initialize render window only once
- The render window can only be initialized only once
per session in the current setup, otherwise there is no
proper binding to the new render window. This can be easily
checked by switching to a new input file in the current
state of the tool. This commit fixes the underlying issue
by only initializing the render window if it was not saved
on the server side.
- Also, there was a non-existent function used for the
determination of the current material. This is fixed as well.