feat: enable vtu file support
- Enables external vtu file support,
see new default file. This is done in
the same framework as for the Exodus
files.
- Improves testing suite: we now test
for different files with different geometry
types: exo, vtu, legacy. Also, the vtu path
is tested for, which more reliably verifies
if the geometry of the yaml file was initialized
correctly.
- Improves rendering of design conditions:
threshold was replaced by more reliable point
filtering. Hence, only the relevant points are
included, and not all cells containing these
relevant points.